requirements.txt
flask 
waitress
psycopg2-binary
kafka
urllib3 == 1.26.13
chardet
python-dotenv
attacut

# sudo apt-get install libpq5=12.12-0ubuntu0.20.04.1 && sudo apt-get install libpq-dev
